QiskitRuntime.jl

Warning

This documentation is not for the standard client qiskit-ibm-runtime to the Qiskit Runtime REST API

To find information on the easiest, best, way to use the Qiskit Runtime use this link:

qiskit-ibm-runtime

The documentation you are reading now is only for the highly experimental Julia-language client, not for the Python-langauge client.

Warning

Documentation pages for QiskitRuntime.jl are a WIP.

Contents

The contents are jumbled and out of order. Documenter is difficult.

Development notes

Native Julia types

Precompile

Name spaces

Data structures

Load times and TTFX

Performance

Exceptions

Sum/Algebraic types

Code quality and CI

QiskitRuntime.jl

Contents

Introduction

Jobs

Backends

Accounts

Instances

Index

Requests Layer

Tutorial

Introduction

QiskitRuntime.QiskitRuntime — Module

module QiskitRuntime

QiskitRuntime is a client for the Qiskit Runtime REST API written in the Julia language.

QiskitRuntime is analagous to the Python-language client qiskit-ibm-runtime.

Warning

QiskitRuntime is very new, incomplete, and API-unstable.

QiskitRuntime.jl is completely unsupported. No person or entity is responsible for providing any support to users of this software.

Furthermore, the focus at the moment is more on retreiving data than on submitting workloads.

Environment variables

The following environment variables override defaults.

QISKIT_RUNTIME_CACHE_DIR: The top-level directory where REST responses are cached.

QISKIT_IBM_URL: The url used for authentication (not for REST endpoints).

QISKIT_IBM_CHANNEL:

QISKIT_IBM_INSTANCE: In the form "hub/group/project".

QISKIT_IBM_TOKEN: The authentication token

Jobs

QiskitRuntime.Jobs.cached_job_ids — Method

cached_job_ids()

Return an iterator over JobIds of cached jobs.

Use job_ids to request ids from the REST API.

QiskitRuntime.Jobs.cached_jobs — Method

cached_jobs(; params::Bool=true, results::Bool=true)

Return an iterator over all cached jobs.

params: If true include job input parameters (including the pubs). Otherwise the field params has value nothing.

results: If true get the job results, if available. Otherwise, the field results has value nothing.

QiskitRuntime.Jobs.job — Function

job(job_id::JobId, account=nothing;  params::Bool=true, results::Bool=true, refresh::Bool=false)::RuntimeJob

Return information on job_id.

params: If true include job input parameters (including the PUBs)). Otherwise the field params has value nothing.

results: If true get the job results, if available. Otherwise, the field results has value nothing. Results are retrieved via the function results.

refresh: If true fetch job info and results from the REST API, rather than from cache. This also updates the cache. If refresh is false, then the cache is preferred.

See results

QiskitRuntime.Jobs.job — Function

job(job_in::RuntimeJob, account=nothing;  params::Bool=true, results::Bool=true, refresh::Bool=false)::RuntimeJob

Return information on job_in.

The fields job_in.params and job_in.results may have value nothing. Use this method to return a copy of job_in with one or both of these fields populated, according to the keyword arguments params and results. These keyword arguments are described in job

Note

This function would be more useful if it were optimized to fetch only the needed additional data, copying the rest from jobin. In fact, at present, it constructs the entire RuntimeJob from scratch.

QiskitRuntime.Jobs.job_ids — Function

job_ids(account=nothing)

Return an iterator over JobIds for all jobs.

job_ids first requests all the information on the jobs, then extracts the ids. This function always makes requests to the REST API and does not access the cache.

Use cached_job_ids to get ids only for cached job requests.

QiskitRuntime.Jobs.results — Function

results(job_id, account=nothing; refresh=false)

Return results for job_id.

See job

QiskitRuntime.Jobs.results — Function

results(job::RuntimeJob, account=nothing; refresh=false)

Return results associated with job.

Results already contained in job are returned if it makes sense to do so.

More precisely, if job.results is not nothing and refresh is false, then job.results is returned. If job.results is nothing then results are fetched from the cache or the REST API. If refresh is false then the cache is preferred.

QiskitRuntime.Jobs.user — Function

user(account=nothing)

Return information about the user.

Information includes the user's email and a list of available instances.

Backends

QiskitRuntime.Backends.backend_status — Function

backend_status(backend_name::AbstractString, account=nothing)

Return status information of backend_name.

QiskitRuntime.Backends.backends — Function

backends(account=nothing; pending=false, testing=false, instance=nothing)

Return a list of available backends.

pending: If true, return a list of tuples (name, num_pending_jobs) sorted by num_pending_jobs.

testing: If true include test devices, those that begin with "test_".

Accounts

QiskitRuntime.Accounts.QuantumAccount — Type

struct that represents an account with channel 'ibm_quantum.'"

QiskitRuntime.Accounts.QuantumAccount — Method

QuantumAccount()

Return a structure that contains information necessary for making requests to the REST API. This includes an access token and an instance.

This reads ~/.qiskit/qiskit-ibm.json. Environment variables take precedent. The latter has been implemented but is not tested.

QiskitRuntime.Accounts.UserId — Method

UserId(id_str::AbstractString)

Construct user id from the user id string in REST response.

Instances

QiskitRuntime.Instances.Instance — Method

Instance(instance::AbstractString)

Construct an Instance from a string of the form "hub/group/project".

Index

QiskitRuntime.QiskitRuntime

QiskitRuntime.Requests

QiskitRuntime.Accounts.QuantumAccount

QiskitRuntime.Accounts.QuantumAccount

QiskitRuntime.Accounts.UserId

QiskitRuntime.Instances.Instance

QiskitRuntime.Backends.backend_status

QiskitRuntime.Backends.backends

QiskitRuntime.Jobs.cached_job_ids

QiskitRuntime.Jobs.cached_jobs

QiskitRuntime.Jobs.job

QiskitRuntime.Jobs.job

QiskitRuntime.Jobs.job_ids

QiskitRuntime.Jobs.results

QiskitRuntime.Jobs.results

QiskitRuntime.Jobs.user

QiskitRuntime.Requests.backends

QiskitRuntime.Requests.cached_job_ids

QiskitRuntime.Requests.job

QiskitRuntime.Requests.job_ids

QiskitRuntime.Requests.jobs

QiskitRuntime.Requests.metrics

QiskitRuntime.Requests.results

QiskitRuntime.Requests.transpiled_circuits

QiskitRuntime.Requests.user

QiskitRuntime.Requests.user_instances

QiskitRuntime.Requests.user_jobs

QiskitRuntime.Requests.workloads