Accounts

Credentials file

Danger

Account credentials are saved in plain text, so only do so if you are using a trusted device.

In order to use the Qiskit Runtime service you need an account, and credentials for authentication.

By default, the credentials are saved in the file

$HOME/.qiskit/qiskit-ibm.json

where $HOME is your home directory. QiskitRuntime can read this file, but cannot write it. You can create it with the Python client and/or edit it by hand.

Which account is used for credentials may also be controlled with environment variables

Info

We want to consistently use the term "credentials file" when referring to the file storing credentials for accounts. However, the word "account" is sometimes used. For example, the credentials file is read, and entries decoded into QuantumAccount objects.

The Python code in qiskit-ibm-runtime sometimes uses the word fragment "config" in variables related to the credentials file, and its containing directory. We do not follow this convention.

Functions and types

QiskitRuntime.Accounts.QuantumAccount — Type

QuantumAccount(account_name=nothing)::QuantumAccount

Return a struct with information for making requests to the REST API.

The account argument may be omitted making requests to the server, for example when calling QiskitRuntime.Jobs.job. In these cases, the account will be constructed with the form QuantumAccount().

The following are tried in order, and the first to succeed is returned.

If account_name is not nothing then the account with this name is read from the credentials file.

If the account information is specified in enviroment variables QISKIT_IBM_INSTANCE and QISKIT_IBM_TOKEN, then these are used to construct the QuantumAccount. The credentials file is not read. (And need not exist.)

If the environment variable QISKIT_ACCOUNT_NAME is set, then this account is read from the credentials file.

The default account "default-ibm-quantum" is read from the credentials file.

The credentials file is typically ~/.qiskit/qiskit-ibm.json.

In case the account is constructed from environment variables, the variables QISKIT_IBM_CHANNEL, QISKIT_IBM_AUTH_URL may be set as well to override defaults.

Examples

(The tokens and instances below are invented, and are not related to any real account.)

julia> accts = list_accounts() # List the accounts in the credentials file.
2-element Vector{String}:
 "default-ibm-quantum"
 "qiskit-other"

julia> QuantumAccount(accts[1]) # Get the first account
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  auth_url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("de638e834a4925507bf4181220eff41d116a7baabd9a339cc2e8a3be570f6196cec2f3d77aabc773e1ae1e62ba67b3a5a7d0bbe9de015b89ec65692803c547e9"),
  instance = Instance(hub-one/group-one/project-one),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

julia> QuantumAccount(accts[2]) # Get the second account
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  auth_url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("726f2418e17c5845aad93c8e8a3cccaa642b9473132aaec47da93cb2f619fe3d156fb4962ebf8865bed278970bbddc3db0caf58c429b2fffd8c8c0359d80b43a"),
  instance = Instance(hub-two/group-two/project-two),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

julia> QuantumAccount() # Get the default account, "default-ibm-quantum"
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  auth_url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("de638e834a4925507bf4181220eff41d116a7baabd9a339cc2e8a3be570f6196cec2f3d77aabc773e1ae1e62ba67b3a5a7d0bbe9de015b89ec65692803c547e9"),
  instance = Instance(hub-one/group-one/project-one),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

Using QiskitRuntime.EnvVars.set_env!, we change the default account name with an environment variable. To reduce verbosity, we just show the instance.

julia> set_env!(:QISKIT_ACCOUNT_NAME, "qiskit-other");

julia> QuantumAccount().instance
Instance(hub-two/group-two/project-two)

Here we set the account with environment variables, ignoring the credentials file.

julia> set_env!(:QISKIT_IBM_TOKEN, "f99b2a42c50ef0314434b0a5e60d8eab6b08112dd4259446e7172a34d61c48ea0a1b349e367dd62d7088fc0d0e97948f0d09fe46c4e34dd9550cdee97256e9a8");

julia> set_env!(:QISKIT_IBM_INSTANCE, "a/b/c");

julia> QuantumAccount()
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  auth_url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("f99b2a42c50ef0314434b0a5e60d8eab6b08112dd4259446e7172a34d61c48ea0a1b349e367dd62d7088fc0d0e97948f0d09fe46c4e34dd9550cdee97256e9a8"),
  instance = Instance(a/b/c),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

julia> foreach(k -> set_env!(k, nothing), (:QISKIT_IBM_INSTANCE, :QISKIT_IBM_TOKEN, :QISKIT_ACCOUNT_NAME));

QiskitRuntime.Accounts.QuantumAccount — Type

struct that stores credentials data for an account."

QiskitRuntime.Accounts.all_accounts — Method

all_accounts() :: Vector{QuantumAccount}

Return a list of all QuantumAccounts in the user's credentials file.

QiskitRuntime.Accounts.list_accounts — Method

list_accounts() :: Vector{String}

Return a list of all account names from the user's credentials file.