Environment Variables

QiskitRuntime.EnvVars — Module

module EnvVars

Enforce documentation and a safe interface to the environment variables used by QiskitRuntime.jl

"safe" means that typos in environment variable names will result in immediate run time or compile time errors.

QiskitRuntime.EnvVars.env_vars — Method

env_vars()

Return a Dict of all environment variables used by QiskitRuntime.jl and their values.

If the environment variable is not set then its value is nothing in the returned Dict. Here "not set" means it is not a key in Base.ENV.

Variables:

QISKIT_USER_DIR - The directory storing user data for Qiskit Runtime, including credentials and cache.

QISKIT_ACCOUNT_NAME - The name of the account in the credentials file to use by default.

QISKIT_IBM_INSTANCE - The instance ("hub/group/project") used only when setting the account via environment variables.

QISKIT_IBM_TOKEN - The token, or API key, used only when setting the account via environment variables.

QISKIT_RUNTIME_CACHE_DIR - The full path of the REST API response cache overriding the default HOME/.qiskit/runtime_cache.

QISKIT_IBM_CHANNEL - The channel name used only when setting the account via environment variables. Currently unused by QiskitRuntime.jl

QISKIT_IBM_AUTH_URL - The url used for authentication. Currently unused by QiskitRuntime.jl

QISKIT_IBM_RUNTIME_LOG_LEVEL - Currently unused by QiskitRuntime.jl

QiskitRuntime.EnvVars.get_env — Function

get_env(name::Symbol, default=nothing)

Return the value for environment variable name, or default if name is not present.

Note

An error is thrown if name is not an environment variable used by QiskitRuntime.jl.

QiskitRuntime.EnvVars.set_env! — Method

set_env!(name::Symbol, val::Union{AbstractString, Nothing})

Set the value for environment variable name to val.

If val is nothing, then name is deleted from Base.ENV.

Note

An error is thrown if name is not an environment variable used by QiskitRuntime.jl.