Id Numbers and Tokens

QiskitRuntime.Ids — Module

module Ids

This module implements identification numbers and related objects.

structs implemented here include: JobId, UserId, and Ids.Token.

Note

We may want to expose less from this module than we do at present. For example, generating bogus ids is useful for mocking, but probabaly not for most users.

QiskitRuntime.Ids.JobId — Type

struct JobId

Wrapper type for job ids

Because a session id is the first job id in a batch, JobId is used for session id as well. The data is stored as a string that is validated on construction.

Note that passing validation is a necessary, but not sufficient condition for distinguishing job ids returned by the server. We could investigate the format of job ids and tighten up the validation.

See validate

QiskitRuntime.Ids.Token — Type

struct Token

Wraps authentication tokens.

The server and the web dashboard express tokens as strings of 128 lower-case hexadecimal digits. We store this as eight UInt64s. In particular, the Token is validated upon construction.

Uppercase hexadecimal characters are allowed when parsing.

julia> typeof(Token("884fad8b23e0cb2e19ae5df80aab5003e968ea4e3a69c6efce9a776b26cb157bb7456b189a964bdba89423ecb2e7b4d23c2644a672c9ba6ef2a1551bed5879d3"))
Token

julia> Token("abc123")
ERROR: ArgumentError: Token must be 128 hex digits

julia> Token("ABCDAD8B23E0CB2E19AE5DF80AAB5003E968EA4E3A69C6EFCE9A776B26CB157BB7456B189A964BDBA89423ECB2E7B4D23C2644A672C9BA6EF2A1551BED5879D3")
Token("abcdad8b23e0cb2e19ae5df80aab5003e968ea4e3a69c6efce9a776b26cb157bb7456b189a964bdba89423ecb2e7b4d23c2644a672c9ba6ef2a1551bed5879d3")

QiskitRuntime.Ids.UserId — Type

struct UserId

Wraps a user id.

The server returns 24 hex digits. We encode this as 96 bits in three UInt32s. Thus, validation occurs on construction.

Base.rand — Method

Random.rand(rng::Random.AbstractRNG, ::Random.SamplerType{JobId})

Return a random JobId.

The set of strings sampled is actually larger than that of true job ids. See JobId.

Base.rand — Method

Random.rand(rng::Random.AbstractRNG, ::Random.SamplerType{Token}) =

Generate a random Token.

This might be useful for putting a bogus token in your credentials file.

Base.rand — Method

Random.rand(rng::Random.AbstractRNG, ::Random.SamplerType{UserId}) =

Construct a random UserId.

This is done by generating and wrapping three UInt32s.

Note

We may not want to expose this. It is used for mocking data.

QiskitRuntime.Ids.validate — Method

validate(::Type{JobId}, job_id::AbstractString)

Return true if job_id is a valid job id.

job_id must be a twenty digit string of 0-9 and a-z. There may be other restrictions, but we don't know about them and cannot check them.