Requests Layer

QiskitRuntime.Requests — Module

module Requests

This module manages requests and responses to the REST API.

Several functions in the Requests layer make requests to specific endpoints. For some endpoints, responses are cached as JSON text files. A request to an endpoint with a cache works as follows:

A function requesting a specific endpoint is called. For example job. In this case, a job id is passed as a parameter.

The cached response is looked up by job id. If a cached response is found, it is read, converted to a JSON3.Object, and returned.

If the cached response is not found, a request is made to the endpoint. The response string is written to the cache and is also converted to a JSON3.Object and returned.

If you pass the keyword parameter refresh=true, the above scenario is modified. In this case, the request is sent to the endpoint unconditionally. The cache file, if present, is overwritten with the new response.

Some endpoints have no associated cache. A function in the Requests layer accessing this endpoint simply makes the requests, parses the result, and returns a JSON3.Object.

The following two arguments are common to many functions in the API of the Requests layer.

Arguments

– qaccount::Union{QuantumAccount, Nothing}=nothing: The QuantumAccount to use. If nothing a QuantumAccount is created.

Keyword arguments

– refresh::Bool=false: If false, then a cached response will be preferred. If true, the a request will be made, and the cache will be updated with a fresh response.

Options to requests

Many endpoints in the REST API support options, such as filters. Most of these are not yet implemented, although it is not difficult to do so.

Warning

No measures are taken to protect the cache from corruption in case of errors.

QiskitRuntime.Requests.RuntimeServiceException — Type

RuntimeServiceException{T} <: Exception

The exception thrown when the status returned in the response to a GET or POST request indicates an error.

Fields

status::Int16: the status code returned in the response

body::{<:JSON3.Object}: The body of the response as a JSON3 object

QiskitRuntime.Requests.backends — Function

backends(qaccount=nothing; provider=nothing)::JSON3.Object

Return list of backends available to the configured user and instance.

If provider is :all, then return all backends, even those not available to the instance.

I'm pretty sure that "provider" and "instance" mean the same thing.

endpoint: "backends"

QiskitRuntime.Requests.cached_job_ids — Method

cached_job_ids()::Generator{Vector{String}}

Return an iterator over all the job ids associated with cached job info.

These are the ids of jobs that were fetched via job or jobs.

QiskitRuntime.Requests.job — Function

job(job_id, qaccount=nothing; refresh=false)::JSON3.Object

Retrieve job info for job_id

endpoint: "job/{job_id}"

QiskitRuntime.Requests.job_ids — Function

job_ids(qaccount=nothing; kws...)::Vector{String}

Return all job ids.

The function jobs is called to retrieve the job info and the ids are extracted and returned.

QiskitRuntime.Requests.jobs — Function

jobs(qaccount=nothing; tags=nothing)

Return job info on all jobs.

endpoint: "jobs"

QiskitRuntime.Requests.metrics — Function

metrics(job_id, qaccount=nothing)::JSON3.Object

Return metrics for job_id.

endpoint: "jobs/{job_id}/metrics"

QiskitRuntime.Requests.results — Function

results(job_id, qaccount=nothing; refresh=false)::JSON3.Object

Return the job results for job_id.

endpoint: "jobs/{job_id}/results"

The other kind of data on a job, which we call "job info", is retrieved with job, or jobs

QiskitRuntime.Requests.transpiled_circuits — Function

transpiled_circuits(job_id, qaccount=nothing)::JSON3.Object

Return transpiled circuits for job_id.

endpoint: "jobs/{job_id}/transpiled_circuits"

QiskitRuntime.Requests.user_info — Function

user_info(qaccount=nothing; refresh=false)::JSON3.Object

Get the authenticated user.

The response includes the user's email and the same information returned by user_instances.

endpoint: "users/me"

QiskitRuntime.Requests.user_instances — Function

user_instances(qaccount=nothing)::JSON3.Object

Return a list of instances available to the user.

endpoint: "instances"

QiskitRuntime.Requests.user_jobs — Function

user_jobs(qaccount=nothing)::JSON3.Object

Mysterious alternative to jobs that returns slightly different results.

endpoint: "facade/v1/jobs"

QiskitRuntime.Requests.workloads — Function

workloads(qaccount=nothing; instance=nothing)::JSON3.Object

List user workloads

Compared to jobs, workloads returns a smaller dictionary with less information for each job.

The default filters are different. Not just the limit on the number of returned jobs.

endpoint: "workloads/me"