QiskitRuntime.jl

Warning

This documentation is not for the standard client qiskit-ibm-runtime to the Qiskit Runtime REST API

To find information on the easiest, best, way to use the Qiskit Runtime use this link:

qiskit-ibm-runtime

The documentation you are reading now is only for the highly experimental Julia-language client, not for the Python-langauge client.

Warning

Documentation pages for QiskitRuntime.jl are a WIP.

Contents

The contents are jumbled and out of order. Documenter is difficult.

Development notes

Native Julia types

Precompile

Name spaces

Data structures

Load times and TTFX

Performance

Exceptions

Sum/Algebraic types

Code quality and CI

QiskitRuntime.jl

Contents

Introduction

Jobs

Backends

Accounts

Instances

PUBs

Circuits

Index

Requests Layer

Tutorial

Introduction

QiskitRuntime.QiskitRuntime — Module

module QiskitRuntime

QiskitRuntime is a client for the Qiskit Runtime REST API written in the Julia language.

QiskitRuntime is analagous to the Python-language client qiskit-ibm-runtime.

Warning

QiskitRuntime is very new, incomplete, and API-unstable.

QiskitRuntime.jl is completely unsupported. No person or entity is responsible for providing any support to users of this software.

Furthermore, the focus at the moment is more on retreiving data than on submitting workloads.

Environment variables

The following environment variables override defaults.

QISKIT_RUNTIME_CACHE_DIR: The top-level directory where REST responses are cached.

QISKIT_IBM_AUTH_URL: The url used for authentication (not for REST endpoints).

QISKIT_IBM_CHANNEL:

QISKIT_IBM_INSTANCE: In the form "hub/group/project".

QISKIT_IBM_TOKEN: The authentication token

Jobs

QiskitRuntime.Jobs.Cancelled — Constant

job has been cancelled

QiskitRuntime.Jobs.Done — Constant

job has successfully run

QiskitRuntime.Jobs.Error — Constant

job incurred error

QiskitRuntime.Jobs.Queued — Constant

job is queued

QiskitRuntime.Jobs.Running — Constant

job is actively running

QiskitRuntime.Jobs.JobStatus — Type

JobStatus

Status of a job as reported by a query to the Runtime.

Queued

Running

Done

Error

Cancelled

QiskitRuntime.Jobs.cached_job_ids — Method

cached_job_ids()

Return an iterator over JobIds of cached jobs.

Use job_ids to request ids from the REST API.

QiskitRuntime.Jobs.cached_jobs — Method

cached_jobs(; params::Bool=true, results::Bool=true)

Return an iterator over all cached jobs.

params: If true include job input parameters (including the pubs). Otherwise the field params has value nothing.

results: If true get the job results, if available. Otherwise, the field results has value nothing.

QiskitRuntime.Jobs.job — Function

job(job_in::RuntimeJob, account=nothing;  params::Bool=true, results::Bool=true, refresh::Bool=false)::RuntimeJob

Return information on job_in.

The fields job_in.params and job_in.results may have value nothing. Use this method to return a copy of job_in with one or both of these fields populated, according to the keyword arguments params and results. These keyword arguments are described in job

Note

This function would be more useful if it were optimized to fetch only the needed additional data, copying the rest from jobin. In fact, at present, it constructs the entire RuntimeJob from scratch.

QiskitRuntime.Jobs.job — Function

job(job_id::JobId, account=nothing;  params::Bool=true, results::Bool=true, refresh::Bool=false)::RuntimeJob

Return information on job_id.

params: If true include job input parameters (including the PUBs)). Otherwise the field params has value nothing.

results: If true get the job results, if available. Otherwise, the field results has value nothing. Results are retrieved via the function results.

refresh: If true fetch job info and results from the REST API, rather than from cache. This also updates the cache. If refresh is false, then the cache is preferred.

See results

QiskitRuntime.Jobs.job_ids — Function

job_ids(account=nothing)

Return an iterator over JobIds for all jobs.

job_ids first requests all the information on the jobs, then extracts the ids. This function always makes requests to the REST API and does not access the cache.

Use cached_job_ids to get ids only for cached job requests.

QiskitRuntime.Jobs.results — Function

results(job_id, account=nothing; refresh=false)

Return results for job_id.

See job

QiskitRuntime.Jobs.results — Function

results(job::RuntimeJob, account=nothing; refresh=false)

Return results associated with job.

Results already contained in job are returned if it makes sense to do so.

More precisely, if job.results is not nothing and refresh is false, then job.results is returned. If job.results is nothing then results are fetched from the cache or the REST API. If refresh is false then the cache is preferred.

QiskitRuntime.Jobs.run_job — Function

run_job(backend_name::AbstractString, pubs, qaccount=nothing)

Run pubs on backend_name.

Parameters for controlling error mitigation are not yet supported.

QiskitRuntime.Jobs.user — Function

user(account=nothing)

Return information about the user.

Information includes the user's email and a list of available instances.

Backends

QiskitRuntime.Backends.backend_status — Function

backend_status(backend_name::AbstractString, account=nothing)

Return status information of backend_name.

QiskitRuntime.Backends.backends — Function

backends(account=nothing; pending=false, testing=false, instance=nothing)

Return a list of available backends.

pending: If true, return a list of tuples (name, num_pending_jobs) sorted by num_pending_jobs.

testing: If true include test devices, those that begin with "test_".

Accounts

QiskitRuntime.Accounts.QuantumAccount — Type

QuantumAccount(name=nothing)::QuantumAccount

Return a struct with information for making requests to the REST API.

The account argument may be omitted making requests to the server, for example when calling QiskitRuntime.Jobs.job. In these cases, the account will be constructed with the form QuantumAccount().

The following are tried in order, and the first to succeed is returned.

If name is not nothing then the account with this name is read from the configuration file.

If the account information is specified in enviroment variables QISKIT_IBM_INSTANCE and QISKIT_IBM_TOKEN, then these are used to construct the QuantumAccount. The configuration file is not read. (And need not exist.)

If the environment variable QISKIT_ACCOUNT_NAME is set, then this account is read from the configuration file.

The default account "default-ibm-quantum" is read from the configuration file.

The configuration file is typically ~/.qiskit/qiskit-ibm.json.

In case the account is constructed from environment variables, the variables QISKIT_IBM_CHANNEL, QISKIT_IBM_AUTH_URL may be set as well to override defaults.

Examples

(The tokens and instances below are invented, and are not related to any real account.)

julia> accts = list_accounts() # List the accounts in the config file.
2-element Vector{String}:
 "default-ibm-quantum"
 "qiskit-other"

julia> QuantumAccount(accts[1]) # Get the first account
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("de638e834a4925507bf4181220eff41d116a7baabd9a339cc2e8a3be570f6196cec2f3d77aabc773e1ae1e62ba67b3a5a7d0bbe9de015b89ec65692803c547e9"),
  instance = Instance(hub-one/group-one/project-one),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

julia> QuantumAccount(accts[2]) # Get the second account
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("726f2418e17c5845aad93c8e8a3cccaa642b9473132aaec47da93cb2f619fe3d156fb4962ebf8865bed278970bbddc3db0caf58c429b2fffd8c8c0359d80b43a"),
  instance = Instance(hub-two/group-two/project-two),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

julia> QuantumAccount() # Get the default account, "default-ibm-quantum"
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("de638e834a4925507bf4181220eff41d116a7baabd9a339cc2e8a3be570f6196cec2f3d77aabc773e1ae1e62ba67b3a5a7d0bbe9de015b89ec65692803c547e9"),
  instance = Instance(hub-one/group-one/project-one),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

We change the default account name with an environment variable. To reduce verbosity, we just show the instance.

julia> ENV["QISKIT_ACCOUNT_NAME"] = "qiskit-other";

julia> QuantumAccount().instance
Instance(hub-two/group-two/project-two)

Here we set the account with environment variables, ignoring the configuration file.

julia> ENV["QISKIT_IBM_TOKEN"] = "f99b2a42c50ef0314434b0a5e60d8eab6b08112dd4259446e7172a34d61c48ea0a1b349e367dd62d7088fc0d0e97948f0d09fe46c4e34dd9550cdee97256e9a8";

julia> ENV["QISKIT_IBM_INSTANCE"] = "a/b/c";

julia> QuantumAccount()
QuantumAccount{Nothing}(
  channel = "ibm_quantum",
  url = "https://auth.quantum-computing.ibm.com/api",
  token = Token("f99b2a42c50ef0314434b0a5e60d8eab6b08112dd4259446e7172a34d61c48ea0a1b349e367dd62d7088fc0d0e97948f0d09fe46c4e34dd9550cdee97256e9a8"),
  instance = Instance(a/b/c),
  private_endpoint = false,
  verify = false,
  proxies = nothing
)

julia> foreach(k -> delete!(ENV, k), ("QISKIT_IBM_INSTANCE", "QISKIT_IBM_TOKEN", "QISKIT_ACCOUNT_NAME"));

QiskitRuntime.Accounts.QuantumAccount — Type

struct that represents an account with channel 'ibm_quantum.'"

Instances

QiskitRuntime.Instances.Instance — Method

Instance(instance::AbstractString)

Construct an Instance from a string of the form "hub/group/project".

julia> inst = Instance("a_hub/a_group/a_project")
Instance(a_hub/a_group/a_project)

julia> inst.project
"a_project"

julia> Instance("a_hub/a_group/a_project/")
ERROR: ArgumentError: Expecting three parts separated by '/'. Got 4 parts

PUBs

QiskitRuntime.PUBs.EstimatorPUB — Type

EstimatorPUB{CircT, ParamsT}

A PUB for the Estimator primitive.

QiskitRuntime.PUBs.SamplerPUB — Type

SamplerPUB{CircT, ParamsT}

A PUB for the Sampler primitive.

Circuits

QiskitRuntime.Circuits.CircuitString — Type

struct CircuitString

Serialized, compressed, encoded data of Python QuantumCircuit type

The QuantumCircuit was serialized as qpy, then compressed with zlib, then base64 encoded.

QiskitRuntime.Circuits.QASMString — Type

QASMString

OpenQASM 3 program as a String.

Index

QiskitRuntime.QiskitRuntime

QiskitRuntime.Requests

QiskitRuntime.Jobs.Cancelled

QiskitRuntime.Jobs.Done

QiskitRuntime.Jobs.Error

QiskitRuntime.Jobs.Queued

QiskitRuntime.Jobs.Running

QiskitRuntime.Accounts.QuantumAccount

QiskitRuntime.Accounts.QuantumAccount

QiskitRuntime.Circuits.CircuitString

QiskitRuntime.Circuits.QASMString

QiskitRuntime.Instances.Instance

QiskitRuntime.Jobs.JobStatus

QiskitRuntime.PUBs.EstimatorPUB

QiskitRuntime.PUBs.SamplerPUB

QiskitRuntime.Backends.backend_status

QiskitRuntime.Backends.backends

QiskitRuntime.Jobs.cached_job_ids

QiskitRuntime.Jobs.cached_jobs

QiskitRuntime.Jobs.job

QiskitRuntime.Jobs.job

QiskitRuntime.Jobs.job_ids

QiskitRuntime.Jobs.results

QiskitRuntime.Jobs.results

QiskitRuntime.Jobs.run_job

QiskitRuntime.Jobs.user

QiskitRuntime.Requests.backends

QiskitRuntime.Requests.cached_job_ids

QiskitRuntime.Requests.job

QiskitRuntime.Requests.job_ids

QiskitRuntime.Requests.jobs

QiskitRuntime.Requests.metrics

QiskitRuntime.Requests.results

QiskitRuntime.Requests.transpiled_circuits

QiskitRuntime.Requests.user

QiskitRuntime.Requests.user_instances

QiskitRuntime.Requests.user_jobs

QiskitRuntime.Requests.workloads