QiskitRuntime.jl

Warning

This documentation is not for the standard client qiskit-ibm-runtime to the Qiskit Runtime REST API

To find information on the easiest, best, way to use the Qiskit Runtime use this link:

qiskit-ibm-runtime

The documentation you are reading now is only for the highly experimental Julia-language client, not for the Python-langauge client.

Warning

Documentation pages for QiskitRuntime.jl are a WIP.

Contents

See also the Development notes.

QiskitRuntime.jl

Contents

Introduction

Jobs

Backends

Instances

PUBs

Circuits

ExtraEnvs

Tutorial

Accounts

Credentials file

Functions and types

Environment Variables

Requests Layer

Id Numbers and Tokens

Index

Introduction

QiskitRuntime.QiskitRuntime — Module

module QiskitRuntime

QiskitRuntime is a client for the Qiskit Runtime REST API written in the Julia language.

QiskitRuntime is analagous to the Python-language client qiskit-ibm-runtime.

Warning

QiskitRuntime is very new, incomplete, and API-unstable.

QiskitRuntime.jl is completely unsupported. No person or entity is responsible for providing any support to users of this software.

Accounts

Many functions, such as job, jobs, user take an optional argument account. If account is omitted, then information will be taken from the user's credentials file ~/.qisit/qiskit-ibm.json, or environment variables. The environment variables will be preferred. See Accounts.QuantumAccount.

Layers

There are more or less two layers: An interface to the REST API, and a layer on top that returns data of native and custom Julia types.

Caching

Caching is done at the level of entire REST API responses.

Reponses from several endpoints are cached automatically. They can be updated with refresh=true. For example Requests.job(job_id; refresh=true).

Functions in the upper layer also take the keyword argument refresh and pass it to the Requests layer. For example Jobs.job(job_id; refresh=true).

Caching is done by dumping the REST responses via JSON3 in ~/.qiskit/runtime_cache/.

Jobs

QiskitRuntime.Jobs.Cancelled — Constant

job has been cancelled

QiskitRuntime.Jobs.Done — Constant

job has successfully run

QiskitRuntime.Jobs.Error — Constant

job incurred error

QiskitRuntime.Jobs.Queued — Constant

job is queued

QiskitRuntime.Jobs.Running — Constant

job is actively running

QiskitRuntime.Jobs.JobStatus — Type

JobStatus

Status of a job as reported by a query to the Runtime.

Queued

Running

Done

Error

Cancelled

QiskitRuntime.Jobs.cached_job_ids — Method

cached_job_ids()

Return an iterator over JobIds of cached jobs.

Use job_ids to request ids from the REST API.

QiskitRuntime.Jobs.cached_jobs — Method

cached_jobs(; params::Bool=true, results::Bool=true)

Return an iterator over all cached jobs.

params: If true include job input parameters (including the pubs). Otherwise the field params has value nothing.

results: If true get the job results, if available. Otherwise, the field results has value nothing.

QiskitRuntime.Jobs.job — Function

job(job_id::JobId, account=nothing;  params::Bool=true, results::Bool=true, refresh::Bool=false)::RuntimeJob

Return information on job_id.

params: If true include job input parameters (including the PUBs)). Otherwise the field params has value nothing.

results: If true get the job results, if available. Otherwise, the field results has value nothing. Results are retrieved via the function results.

refresh: If true fetch job info and results from the REST API, rather than from cache. This also updates the cache. If refresh is false, then the cache is preferred.

See results

QiskitRuntime.Jobs.job — Function

job(job_in::RuntimeJob, account=nothing;  params::Bool=true, results::Bool=true, refresh::Bool=false)::RuntimeJob

Return information on job_in.

The fields job_in.params and job_in.results may have value nothing. Use this method to return a copy of job_in with one or both of these fields populated, according to the keyword arguments params and results. These keyword arguments are described in job

Note

This function would be more useful if it were optimized to fetch only the needed additional data, copying the rest from jobin. In fact, at present, it constructs the entire RuntimeJob from scratch.

QiskitRuntime.Jobs.job_ids — Function

job_ids(account=nothing)

Return an iterator over JobIds for all jobs.

job_ids first requests all the information on the jobs, then extracts the ids. This function always makes requests to the REST API and does not access the cache.

Use cached_job_ids to get ids only for cached job requests.

QiskitRuntime.Jobs.results — Function

results(job_id, account=nothing; refresh=false)

Return results for job_id.

See job

QiskitRuntime.Jobs.results — Function

results(job::RuntimeJob, account=nothing; refresh=false)

Return results associated with job.

Results already contained in job are returned if it makes sense to do so.

More precisely, if job.results is not nothing and refresh is false, then job.results is returned. If job.results is nothing then results are fetched from the cache or the REST API. If refresh is false then the cache is preferred.

QiskitRuntime.Jobs.run_job — Function

run_job(backend_name::AbstractString, pubs::AbstractVector{<:AbstractPUB}, qaccount=nothing)

Run pubs on device backend_name.

Parameters for controlling error mitigation are not yet supported.

See Accounts.QuantumAccount, PUBs.EstimatorPUB, PUBs.SamplerPUB, Backends.backends.

QiskitRuntime.Jobs.user_info — Function

user_info(account=nothing; refresh=false)::UserInfo

Return information about the user.

Information includes the user's email and a list of available instances.

Note

The user info returned by the server is determined by the authentication token. Unlike other cached information, we do not key the cache by this information in order to avoid writing the token as plain text. So if you have more than one account, each with an associated token, the cache will not distinguish them. If you switch accounts and tokens, you should pass refresh=true.

If you merely generate and use a new token for a single account, you do not need to refresh the cache.

Backends

QiskitRuntime.Backends.backend_status — Function

backend_status(backend_name::AbstractString, account=nothing)

Return status information of backend_name.

QiskitRuntime.Backends.backends — Function

backends(account=nothing; pending=false, testing=false, instance=nothing)

Return a list of available backends.

pending: If true, return a list of tuples (name, num_pending_jobs) sorted by num_pending_jobs.

testing: If true include test devices, those that begin with "test_".

Instances

QiskitRuntime.Instances.Instance — Method

Instance(instance::AbstractString)

Construct an Instance from a string of the form "hub/group/project".

julia> inst = Instance("a_hub/a_group/a_project")
Instance(a_hub/a_group/a_project)

julia> inst.project
"a_project"

julia> Instance("a_hub/a_group/a_project/")
ERROR: ArgumentError: Expecting three parts separated by '/'. Got 4 parts

PUBs

QiskitRuntime.PUBs.EstimatorPUB — Type

EstimatorPUB{CircT, ParamsT}

A PUB for the Estimator primitive.

QiskitRuntime.PUBs.SamplerPUB — Type

SamplerPUB{CircT, ParamsT}

A PUB for the Sampler primitive.

Circuits

QiskitRuntime.Circuits.CircuitString — Type

struct CircuitString

Serialized, compressed, encoded data of Python QuantumCircuit type

The QuantumCircuit was serialized as qpy, then compressed with zlib, then base64 encoded.

QiskitRuntime.Circuits.QASMString — Type

QASMString

OpenQASM 3 program as a String.

ExtraEnvs

QiskitRuntime.ExtraEnvs — Module

module ExtraEnvs

Contains facilites for managing an environment of extra (auxilliary) packages.

The module defines ExtraEnv and implements functions that use data therein to manipulate the stack of environements (Base.LOAD_PATH).

There are some functions missing. For example, for adding and removing packages from the ExtraEnv. This can be done manually, but is would be a bit fiddle. These functions might be added later.

QiskitRuntime.ExtraEnvs.ExtraEnv — Type

struct ExtraEnv

Data for defining an "extra" shared environment.

When working with a particular package, you might frequently use certain other packages that are not in the dependencies of the particular one. ExtraEnv is meant to help manage these other packages. You can do this by hand as well. But it's sometimes not easy to remember what you want to do and exactly how to do it.

ExtraEnv helps you create a shared environment and make sure it is in your stack of environments so that its packages are visibile when it is not the active environment. You are not meant to do work with the environment in ExtraEnv activated.

The most important function for creating an ExtraEnv and making it visible is ensure_in_stack.

Fields

name::String: the name of the extra environment

packages::Vector{Symbol}: A list of packages to use to initialize the environment.

See create_env, ensure_in_stack, env_exists, activate_env, is_in_stack, delete_from_stack!.

Examples

julia> ExtraEnv("an_extra_env", [:Example])
ExtraEnv("an_extra_env", [:Example])

julia> ExtraEnv("@an_extra_env", [:Example])
ExtraEnv("an_extra_env", [:Example])

QiskitRuntime.ExtraEnvs.ExtraEnv — Method

ExtraEnv(env_name::AbstractString)

Initialize an ExtraEnv with an empty list of packages.

Examples

julia> ExtraEnv("an_extra_env")
ExtraEnv("an_extra_env", Symbol[])

julia> ExtraEnv("@an_extra_env")
ExtraEnv("an_extra_env", Symbol[])

QiskitRuntime.ExtraEnvs.activate_env — Method

activate_env(env::ExtraEnv)

Activate the shared environment env.name.

You might want to do this to add or remove packages from the environment. But it is not necessary to activate it to use it.

See ExtraEnv, create_env, ensure_in_stack, env_exists.

QiskitRuntime.ExtraEnvs.create_env — Method

create_env(env::ExtraEnv)

Create a shared environment named env.name with packages env.packages.

If the shared environment env.name already exists, an error is thrown.

See ExtraEnv.

QiskitRuntime.ExtraEnvs.delete_from_stack! — Method

delete_from_stack!(env::ExtraEnv)

Delete environment env.name wherever it occurs in the stack Base.LOAD_PATH.

See ExtraEnv.

QiskitRuntime.ExtraEnvs.delete_from_stack! — Method

delete_from_stack!(env_name::Union{AbstractString, Symbol})

Delete environment env_name wherever it occurs in the stack Base.LOAD_PATH.

See ExtraEnv.

QiskitRuntime.ExtraEnvs.ensure_in_stack — Method

ensure_in_stack(env_name::AbstractString, env_packages::AbstractVector)::ExtraEnv

Create env = ExtraEnv(env_name, env_packages), run ensure_in_stack(env) and return env.

Elements of env_packages should be AbstractStrings or Symbols.

See ExtraEnv.

Examples

julia> ensure_in_stack("my_extra_env", [:PackageA, :PackageB]);

QiskitRuntime.ExtraEnvs.ensure_in_stack — Method

ensure_in_stack(env::ExtraEnv)::ExtraEnv

Ensure that a shared environment env.name with env.packages is in your stack.

Recall that the environment stack is Base.LOAD_PATH.

If env.name does not name a shared environment, create it and add env.packages. Furthermore, if env.name is not in the stack, add it to the stack.

After this runs, the packages env.packages should be available in whatever project is active.

See ExtraEnv.

QiskitRuntime.ExtraEnvs.env_exists — Method

env_exists(env_name::AbstractString)::Bool

Return true if the shared environment env_name already exists.

env_name may begin with "@" or not.

This environment might be activated via Pkg.activate(env_name) If done from the pkg repl, the name must start with '@'.

QiskitRuntime.ExtraEnvs.env_exists — Method

env_exists(env::ExtraEnv)::Bool

Return true if the shared environment in env already exists.

See ExtraEnv.

QiskitRuntime.ExtraEnvs.is_in_stack — Method

is_in_stack(env_name::AbstractString)::Bool

Return true if the shared environment env_name is in the environment stack.

If env_name does not start with '@', it is prepended. The stack is Base.LOAD_PATH.

QiskitRuntime.ExtraEnvs.is_in_stack — Method

is_in_stack(env::ExtraEnv)::Bool

Return true if the shared environment env.name is in the environment stack.

See ExtraEnv.